+++ test/index.test.ts
@@ -1,29 +1,45 @@
 /* eslint-disable @typescript-eslint/no-extraneous-class,@typescript-eslint/explicit-member-accessibility,@typescript-eslint/no-empty-function */
 import os from 'os';
-import { TempDir } from './util';
 import {
   cosmiconfig as cosmiconfigModule,
   cosmiconfigSync as cosmiconfigSyncModule,
   defaultLoaders,
   LoaderSync,
 } from '../src';
-import { Loaders } from '../src/types';
+import { ExplorerOptions, ExplorerOptionsSync, Loaders } from '../src/types';
+import { TempDir } from './util';
 
 const temp = new TempDir();
 
 function getLoaderFunctionsByName(loaders: Loaders) {
-  const loaderFunctionsByName = Object.entries(loaders).reduce(
-    (acc, [extension, loader]) => {
-      return {
-        ...acc,
-        [extension]: loader.name,
-      };
-    },
-    {},
+  return Object.fromEntries(
+    Object.entries(loaders).map(([extension, loader]) => [
+      extension,
+      loader.name,
+    ]),
   );
-
-  return loaderFunctionsByName;
 }
+
+const checkConfigResult = (
+  mock: jest.SpyInstance,
+  instanceNum: number,
+  expectedLoaderNames: Record<string, string>,
+  expectedExplorerOptions: Omit<
+    ExplorerOptions & ExplorerOptionsSync,
+    'transform' | 'loaders'
+  >,
+) => {
+  const instanceIndex = instanceNum - 1;
+  expect(mock.mock.calls.length).toEqual(instanceNum);
+  expect(mock.mock.calls[instanceIndex].length).toEqual(1);
+
+  const { transform, loaders, ...explorerOptions } =
+    mock.mock.calls[instanceIndex][0];
+  expect(transform.name).toBe('identity');
+  const loaderFunctionsByName = getLoaderFunctionsByName(loaders);
+  expect(loaderFunctionsByName).toEqual(expectedLoaderNames);
+  expect(explorerOptions).toEqual(expectedExplorerOptions);
+};
 
 let cosmiconfig: typeof cosmiconfigModule;
 let cosmiconfigSync: typeof cosmiconfigSyncModule;
@@ -74,55 +90,58 @@ describe('cosmiconfig', () => {
   });
 
   describe('creates explorer with default options if not specified', () => {
-    const checkResult = (mock: jest.SpyInstance) => {
-      expect(mock.mock.calls.length).toEqual(1);
-      expect(mock.mock.calls[0].length).toEqual(1);
+    const expectedLoaderNames = {
+      '.cjs': 'loadJs',
+      '.js': 'loadJs',
+      '.json': 'loadJson',
+      '.yaml': 'loadYaml',
+      '.yml': 'loadYaml',
+      noExt: 'loadYaml',
+    };
 
-      const { transform, loaders, ...explorerOptions } = mock.mock.calls[0][0];
-      expect(transform.name).toBe('identity');
-      const loaderFunctionsByName = getLoaderFunctionsByName(loaders);
-      expect(loaderFunctionsByName).toEqual({
-        '.cjs': 'loadJs',
-        '.js': 'loadJs',
-        '.json': 'loadJson',
-        '.yaml': 'loadYaml',
-        '.yml': 'loadYaml',
-        noExt: 'loadYaml',
-      });
-
-      expect(explorerOptions).toEqual({
-        packageProp: moduleName,
-        searchPlaces: [
-          'package.json',
-          `.${moduleName}rc`,
-          `.${moduleName}rc.json`,
-          `.${moduleName}rc.yaml`,
-          `.${moduleName}rc.yml`,
-          `.${moduleName}rc.js`,
-          `.${moduleName}rc.cjs`,
-          `.config/${moduleName}rc`,
-          `.config/${moduleName}rc.json`,
-          `.config/${moduleName}rc.yaml`,
-          `.config/${moduleName}rc.yml`,
-          `.config/${moduleName}rc.js`,
-          `.config/${moduleName}rc.cjs`,
-          `${moduleName}.config.js`,
-          `${moduleName}.config.cjs`,
-        ],
-        ignoreEmptySearchPlaces: true,
-        stopDir: os.homedir(),
-        cache: true,
-      });
+    const expectedExplorerOptions = {
+      packageProp: moduleName,
+      searchPlaces: [
+        'package.json',
+        `.${moduleName}rc`,
+        `.${moduleName}rc.json`,
+        `.${moduleName}rc.yaml`,
+        `.${moduleName}rc.yml`,
+        `.${moduleName}rc.js`,
+        `.${moduleName}rc.cjs`,
+        `.config/${moduleName}rc`,
+        `.config/${moduleName}rc.json`,
+        `.config/${moduleName}rc.yaml`,
+        `.config/${moduleName}rc.yml`,
+        `.config/${moduleName}rc.js`,
+        `.config/${moduleName}rc.cjs`,
+        `${moduleName}.config.js`,
+        `${moduleName}.config.cjs`,
+      ],
+      ignoreEmptySearchPlaces: true,
+      stopDir: os.homedir(),
+      cache: true,
+      metaConfigFilePath: null,
     };
 
     test('async', () => {
       cosmiconfig(moduleName);
-      checkResult(createExplorerMock);
+      checkConfigResult(
+        createExplorerMock,
+        1,
+        expectedLoaderNames,
+        expectedExplorerOptions,
+      );
     });
 
     test('sync', () => {
       cosmiconfigSync(moduleName);
-      checkResult(createExplorerSyncMock);
+      checkConfigResult(
+        createExplorerSyncMock,
+        2,
+        expectedLoaderNames,
+        expectedExplorerOptions,
+      );
     });
   });
 
@@ -171,39 +190,42 @@ describe('cosmiconfig', () => {
       },
     };
 
-    const checkResult = (mock: jest.SpyInstance) => {
-      expect(mock.mock.calls.length).toEqual(1);
-      expect(mock.mock.calls[0].length).toEqual(1);
-      const { transform, loaders, ...explorerOptions } = mock.mock.calls[0][0];
+    const expectedLoaderNames = {
+      '.cjs': 'jsLoader',
+      '.js': 'jsLoader',
+      '.json': 'jsonLoader',
+      '.yaml': 'yamlLoader',
+      '.yml': 'loadYaml',
+      noExt: 'noExtLoader',
+    };
 
-      expect(transform.name).toBe('identity');
-      const loaderFunctionsByName = getLoaderFunctionsByName(loaders);
-      expect(loaderFunctionsByName).toEqual({
-        '.cjs': 'jsLoader',
-        '.js': 'jsLoader',
-        '.json': 'jsonLoader',
-        '.yaml': 'yamlLoader',
-        '.yml': 'loadYaml',
-        noExt: 'noExtLoader',
-      });
-
-      expect(explorerOptions).toEqual({
-        packageProp: 'wildandfree',
-        searchPlaces: ['.foorc.json', 'wildandfree.js'],
-        ignoreEmptySearchPlaces: false,
-        stopDir: __dirname,
-        cache: false,
-      });
+    const expectedExplorerOptions = {
+      packageProp: 'wildandfree',
+      searchPlaces: ['.foorc.json', 'wildandfree.js'],
+      ignoreEmptySearchPlaces: false,
+      stopDir: __dirname,
+      cache: false,
+      metaConfigFilePath: null,
     };
 
     test('async', () => {
       cosmiconfig(moduleName, options);
-      checkResult(createExplorerMock);
+      checkConfigResult(
+        createExplorerMock,
+        1,
+        expectedLoaderNames,
+        expectedExplorerOptions,
+      );
     });
 
     test('sync', () => {
       cosmiconfigSync(moduleName, options);
-      checkResult(createExplorerSyncMock);
+      checkConfigResult(
+        createExplorerSyncMock,
+        2,
+        expectedLoaderNames,
+        expectedExplorerOptions,
+      );
     });
   });