+++ test/successful-directories.test.ts
@@ -67,16 +67,18 @@ describe('finds rc file in third searched dir, with a package.json lacking prop'
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -120,16 +122,18 @@ describe('finds package.json prop in second searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -181,14 +185,16 @@ describe('finds package.json with nested packageProp in second searched dir', ()
   };
 
   test('async', async () => {
+    const explorer = cosmiconfig('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFile');
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
+    const explorer = cosmiconfigSync('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFileSync');
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -231,16 +237,18 @@ describe('finds JS file in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -284,16 +292,18 @@ describe('finds CJS file in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -328,16 +338,18 @@ describe('finds .foorc.js file in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -373,16 +385,18 @@ describe('finds .foorc.cjs file in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -416,16 +430,18 @@ describe("finds foorc file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -460,16 +476,18 @@ describe("finds foorc.json file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -505,16 +523,18 @@ describe("finds foorc.yaml file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -551,16 +571,18 @@ describe("finds foorc.yml file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -601,16 +623,18 @@ describe("finds foorc.js file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -652,16 +676,18 @@ describe("finds foorc.cjs file in first searched dir's .config subdir", () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -705,16 +731,18 @@ describe('skips over empty file to find JS file in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -747,16 +775,18 @@ describe('finds package.json in second dir searched, with alternate names', () =
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -792,15 +822,16 @@ describe('finds rc file in third searched dir, skipping packageProp, parsing ext
   };
 
   test('async', async () => {
+    const explorer = cosmiconfig('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFile');
-
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
+    const explorer = cosmiconfigSync('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFileSync');
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -832,16 +863,18 @@ describe('finds package.json file in second searched dir, skipping JS and RC fil
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -886,16 +919,18 @@ describe('finds .foorc.json in second searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -926,16 +961,18 @@ describe('finds .foorc.yaml in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -967,16 +1004,18 @@ describe('finds .foorc.yml in first searched dir', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1028,15 +1067,16 @@ describe('adding myfooconfig.js to searchPlaces, finds it in first searched dir'
   };
 
   test('async', async () => {
+    const explorer = cosmiconfig('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFile');
-
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
+    const explorer = cosmiconfigSync('foo', explorerOptions);
     const readFileSpy = jest.spyOn(fs, 'readFileSync');
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1101,16 +1141,18 @@ describe('finds JS file traversing from cwd', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search();
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search();
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search();
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search();
     checkResult(readFileSpy, result);
   });
 });
@@ -1150,16 +1192,18 @@ describe('searchPlaces can include subdirectories', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1196,16 +1240,18 @@ describe('directories with the same name as a search place are not treated as fi
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1263,16 +1309,18 @@ describe('custom loaders allow non-default file types', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1330,16 +1378,18 @@ describe('adding custom loaders allows for default and non-default file types', 
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1387,16 +1437,18 @@ describe('defaults loaders can be overridden', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1432,31 +1484,32 @@ describe('custom loaders can be async', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
     const explorerOptions = {
       ...baseOptions,
       loaders: {
         '.things': loadThingsAsync,
       },
     };
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     expect(loadThingsSync).not.toHaveBeenCalled();
     expect(loadThingsAsync).toHaveBeenCalled();
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
-
     const explorerOptions = {
       ...baseOptions,
       loaders: {
         '.things': loadThingsSync,
       },
     };
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     expect(loadThingsSync).toHaveBeenCalled();
     expect(loadThingsAsync).not.toHaveBeenCalled();
     checkResult(readFileSpy, result);
@@ -1494,9 +1547,10 @@ describe('a custom loader entry can include just an async loader', () => {
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });
@@ -1534,16 +1588,18 @@ describe('a custom loader entry can include only a sync loader and work for both
   };
 
   test('async', async () => {
-    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const explorer = cosmiconfig('foo', explorerOptions);
 
-    const result = await cosmiconfig('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFile');
+    const result = await explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 
   test('sync', () => {
-    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const explorer = cosmiconfigSync('foo', explorerOptions);
 
-    const result = cosmiconfigSync('foo', explorerOptions).search(startDir);
+    const readFileSpy = jest.spyOn(fs, 'readFileSync');
+    const result = explorer.search(startDir);
     checkResult(readFileSpy, result);
   });
 });